// resources/js/api/generated/models/operationRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { RegistryCredential } from './registryCredential';

export interface OperationRequest {
  command: string;
  options: string[];
  registry_credentials?: RegistryCredential[];
  services: string[];
}